Added listing for accounts and products

+++ src/main/java/com/example/autoposterbackend/repository/ConfirmationTokenRepository.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Repository
-public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Long> {
+public interface ConfirmationTokenRepository extends JpaRepository<ConfirmationToken, Integer> {
     Optional<ConfirmationToken> findByToken(String token);
 
     @Transactional
